refactor: moved dto content to interface folders
